Updated hooks and created DatabaseValidationSteps where I put all database related methods
