+++ src/test/java/runners/DatabaseRunner.java
@@ -1,0 +1,17 @@
+package runners;
+
+import io.cucumber.junit.Cucumber;
+import io.cucumber.junit.CucumberOptions;
+import org.junit.runner.RunWith;
+
+@RunWith(Cucumber.class)
+@CucumberOptions(
+        features = "src/test/resources/features",
+        glue = "step_definitions",
+        tags = "@DB",
+        plugin = {"html:html-report/report.html","json:target/cucumber.json"},
+        dryRun = false
+)
+
+public class DatabaseRunner {
+}
